+++ packages/core/types/types/index.d.ts
@@ -1,4 +1,4 @@
 export { AddonRegister, AddonGet, Addon, Addons } from "./addon";
 export { Route, RouteDescriptor, ParamParser, ParamParsers, RouteProps, ResponseSetters, ResponseBuilder, EveryMatchFn, InitialMatchFn, ResponseMatchFn } from "./route";
 export { Response, RawParams, Params } from "./response";
-export { CuriRouter, RouterOptions, ResponseHandler, ResponseHandlerProps, RemoveResponseHandler, SideEffect, Cache, Navigation } from "./curi";
+export { CuriRouter, RouterOptions, ResponseHandler, Emitted, RemoveResponseHandler, SideEffect, Cache, Navigation } from "./curi";