+++ packages/side-effects/side-effect-scroll/src/index.ts
@@ -1,7 +1,7 @@
-import { ResponseHandler, ResponseHandlerProps } from "@curi/core";
+import { ResponseHandler, Emitted } from "@curi/core";
 
 function createScrollSideEffect(): ResponseHandler {
-  return function({ response, navigation }: ResponseHandlerProps): void {
+  return function({ response, navigation }: Emitted): void {
     if (navigation.action === "POP") {
       return;
     }